Fix .env loading and json example

+++ src/ACMESkyService/entities/flight_subscription.go
@@ -8,7 +8,7 @@ type CustomerFlightSubscription struct {
 	DateStartISO8601 string `json:"travel_date_start"`
 	// end travel date range in ISO 8601 format (with timezone UTC)
 	DateEndISO8601 string  `json:"travel_date_end"`
-	Budget         float64 `json:"travel_max_price"`
+	Budget         float32 `json:"travel_max_price"`
 	SeatsCount     uint    `json:"travel_seats_count"`
 }
 
@@ -19,8 +19,8 @@ func CustomerFlightSubscriptionFromMap(m map[string]interface{}) CustomerFlightS
 		AirportDestinationID: m["airport_id_destination"].(string),
 		DateStartISO8601:     m["travel_date_start"].(string),
 		DateEndISO8601:       m["travel_date_end"].(string),
-		Budget:               m["travel_max_price"].(float64),
-		SeatsCount:           m["travel_seats_count"].(uint),
+		Budget:               float32(m["travel_max_price"].(float64)),
+		SeatsCount:           uint(m["travel_seats_count"].(float64)),
 	}
 	return v
 }
